requirements.txt
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
connexion==2.14.1
Flask==2.2.3
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.1.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
numpy==1.24.2
packaging==22
paho-mqtt==1.6.1
pyparsing==3.0.9
pyrsistent==0.19.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.42
swagger-ui-bundle==0.0.9
urllib3==1.26.12
Werkzeug==2.2.2
yapf==0.32.0
zipp==3.15.0
attrs==22.1.0
black==23.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
connexion==2.14.1
distlib==0.3.6
filelock==3.9.0
Flask==2.2.3
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.1.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
mypy-extensions==1.0.0
numpy==1.24.2
opencv-python==4.7.0.72
paho-mqtt==1.6.1
pandas==2.0.0
pathspec==0.11.0
platformdirs==3.1.1
protobuf==4.21.12
pyparsing==3.0.9
pyrsistent==0.19.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sklearn==0.0.post4
SQLAlchemy==1.4.42
swagger-ui-bundle==0.0.9
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==1.26.12
virtualenv==20.20.0
Werkzeug==2.2.2
yapf==0.32.0
zipp==3.15.0
